+++ embedded/cult_connect_embbeded/src/air_humidity_sensor.cpp
@@ -1,6 +1,6 @@
 #include "sensors/air_humidity_sensor.h"
 
-AirHumiditySensor::AirHumiditySensor(DHT *dhtSensor, int dhtType, int pin, String id, String type, uint32_t timer, uint32_t repeat, resolution_t resolution) : Sensor(id, type, timer, repeat, resolution)
+AirHumiditySensor::AirHumiditySensor(DHT *dhtSensor, uint8_t dhtType, uint8_t pin, String id, String type, uint32_t timer, uint32_t repeat, resolution_t resolution) : Sensor(id, type, timer, repeat, resolution)
 {
     this->sensor = dhtSensor;
     this->dhtPin = pin;
@@ -15,9 +15,9 @@ float AirHumiditySensor::getMeasure(void)
     float humidity = this->sensor->readHumidity();
     return isnan(humidity) ? -1 : humidity;
 }
-int AirHumiditySensor::getDhtPin(void) { return this->dhtPin; }
+uint8_t AirHumiditySensor::getDhtPin(void) { return this->dhtPin; }
 
-int AirHumiditySensor::getDhtType(void) { return this->dhtType; }
+uint8_t AirHumiditySensor::getDhtType(void) { return this->dhtType; }
 
 String AirHumiditySensor::toString(void)
 {
